Add Lua license

+++ XXTExplorer/Detail/Editor/Settings/XXTEEditorSettingsViewController.m
@@ -273,7 +273,7 @@
     
     XXTEEditorWrapColumnCell *cell11 = [[[NSBundle mainBundle] loadNibNamed:NSStringFromClass([XXTEEditorWrapColumnCell class]) owner:nil options:nil] lastObject];
     cell11.titleLabel.text = NSLocalizedString(@"Word Column", nil);
-    int columnValue = XXTEDefaultsInt(XXTEEditorWrapColumn, 160);
+    NSInteger columnValue = XXTEDefaultsInt(XXTEEditorWrapColumn, 160);
     cell11.valueField.text = [NSString stringWithFormat:@"%d", columnValue];
     cell11.valueField.delegate = self;
     {
